feat(Restaurant): 레스토랑 도메인 생성 Co-authored-by: skiende74 <skiende74@gmail.com>
