+++ tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "incremental": true,
     "target": "ES2020",
+    "types": ["cypress"],
     "lib": ["dom", "dom.iterable", "esnext"],
     "module": "esnext",
     "moduleResolution": "node",
